Added more project: deck_of_cards and a webscraping project
Deck of cards is a CLI program that simulates a deck of cards by shuffling and dealing random cards, drawing cards from the deck, and so on. WebscrapeBlogSite is a program that scrapes a web UI engineer's personal website for her blog posts, and creates a list of her blog topics saved to a csv file.